[feat] 업그레이드 방식 변경 및 데이터 저장

+++ Assets/scrpit/06.21/PlayerData.cs
@@ -5,10 +5,10 @@ public class PlayerData
     public int maxHealth;
     public int money;
 
-    public float bonusSwordDamage;
+    public int bonusSwordDamage;
     public float bonusSwordRange;
-    public float bonusBowDamage;
-    public float bonusPierceCount;
-    public float bonusMaxHealth;
+    public int bonusBowDamage;
+    public int bonusPierceCount;
+    public int bonusMaxHealth;
     public float bonusInvincibleTime;
 }
